Rename LanguageIntegrator project to Languages
