+++ CodeCompletion/CodeCompletion.csproj
@@ -87,9 +87,9 @@
       <Project>{44A01F9E-DCE7-470C-AAE5-C3DE0CCBEE3B}</Project>
       <Name>Errors</Name>
     </ProjectReference>
-    <ProjectReference Include="..\LanguageIntegrator\LanguageIntegrator.csproj">
+    <ProjectReference Include="..\Languages\Languages.csproj">
       <Project>{a48d9069-d569-4110-9252-a10f139b669b}</Project>
-      <Name>LanguageIntegrator</Name>
+      <Name>Languages</Name>
     </ProjectReference>
     <ProjectReference Include="..\ParserTools\ParserTools.csproj">
       <Project>{AF2EFD7B-69DD-4B43-AF65-B59B29349C23}</Project>
