+++ Compiler/Compiler.csproj
@@ -121,9 +121,9 @@
       <Project>{27d9800e-2689-4aa1-a2d6-128e4a9bae98}</Project>
       <Name>LambdaAnySynToSemConverter</Name>
     </ProjectReference>
-    <ProjectReference Include="..\LanguageIntegrator\LanguageIntegrator.csproj">
+    <ProjectReference Include="..\Languages\Languages.csproj">
       <Project>{a48d9069-d569-4110-9252-a10f139b669b}</Project>
-      <Name>LanguageIntegrator</Name>
+      <Name>Languages</Name>
     </ProjectReference>
     <ProjectReference Include="..\Localization\Localization.csproj">
       <Project>{2DE2842F-0912-4251-BC0F-480854C44A13}</Project>
