+++ Languages/Languages.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{A48D9069-D569-4110-9252-A10F139B669B}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>LanguageIntegration</RootNamespace>
-    <AssemblyName>LanguageIntegrator</AssemblyName>
+    <RootNamespace>Languages</RootNamespace>
+    <AssemblyName>Languages</AssemblyName>
     <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
