+++ Languages/Properties/AssemblyInfo.cs
