+++ ReleaseGenerators/sect_Core.nsh
@@ -7,7 +7,7 @@
     File "..\bin\Errors.dll"
     File "..\bin\Localization.dll"
     File "..\bin\NETGenerator.dll"
-    File "..\bin\LanguageIntegrator.dll"
+    File "..\bin\Languages.dll"
     File "..\bin\StringConstants.dll"
     File "..\bin\ParserTools.dll"
     File "..\bin\SemanticTree.dll"
@@ -59,7 +59,7 @@
     ${AddFile} "Errors.dll"
     ${AddFile} "Localization.dll"
     ${AddFile} "NETGenerator.dll"
-    ${AddFile} "LanguageIntegrator.dll"
+    ${AddFile} "Languages.dll"
     ${AddFile} "StringConstants.dll"
     ${AddFile} "ParserTools.dll"
     ${AddFile} "SemanticTree.dll"
