+++ VisualPascalABCNET/DS/VisualEnvironmentCompiler.cs
@@ -7,7 +7,7 @@ using System.Windows.Forms;
 using System.IO;
 using VisualPascalABCPlugins;
 using System.Threading;
-using LanguageIntegration;
+using Languages;
 
 namespace VisualPascalABC
 {
