+++ VisualPascalABCNET/IB/CodeCompletion/TooltipServiceManager.cs
@@ -11,7 +11,7 @@ using ICSharpCode.Core;
 using ICSharpCode.TextEditor;
 using ICSharpCode.TextEditor.Document;
 using ICSharpCode.TextEditor.Gui.CompletionWindow;
-using LanguageIntegration;
+using Languages;
 using PascalABCCompiler.Parsers;
 
 namespace VisualPascalABC
