+++ VisualPascalABCNETLinux/Form1.cs
@@ -681,7 +681,7 @@ namespace VisualPascalABC
                 this.WindowState = FormWindowState.Maximized;
 
             // загрузка всех парсеров и других составляющих языков  EVA
-            LanguageIntegration.LanguageIntegrator.LoadAllLanguages();
+            Languages.LanguageIntegrator.LoadAllLanguages();
 
             ChangedSelectedTab();
             VisualEnvironmentCompiler.ChangeVisualEnvironmentState += new ChangeVisualEnvironmentStateDelegate(VisualEnvironmentCompiler_ChangeVisualEnvironmentState);
