+++ VisualPascalABCNETLinux/IB/CodeCompletion/CodeCompletionParserController.cs
@@ -8,7 +8,7 @@ using System.Windows.Forms;
 using ICSharpCode.TextEditor;
 using ICSharpCode.TextEditor.Gui.CompletionWindow;
 using ICSharpCode.TextEditor.Document;
-using LanguageIntegration;
+using Languages;
 
 namespace VisualPascalABC
 {
