+++ pabcnetc/CommandConsoleCompiler.cs
@@ -304,7 +304,7 @@ namespace PascalABCCompiler
             Console.InputEncoding = System.Text.Encoding.UTF8;
 
             // загрузка всех парсеров и других составляющих языков  EVA
-            LanguageIntegration.LanguageIntegrator.LoadAllLanguages();
+            Languages.LanguageIntegrator.LoadAllLanguages();
             
             LoadCompiler();
