+++ pabcnetc/PABCNETC.csproj
@@ -117,13 +117,13 @@
       <Project>{44A01F9E-DCE7-470C-AAE5-C3DE0CCBEE3B}</Project>
       <Name>Errors</Name>
     </ProjectReference>
+    <ProjectReference Include="..\Languages\Languages.csproj">
+      <Project>{a48d9069-d569-4110-9252-a10f139b669b}</Project>
+      <Name>Languages</Name>
+    </ProjectReference>
     <ProjectReference Include="..\Localization\Localization.csproj">
       <Project>{2DE2842F-0912-4251-BC0F-480854C44A13}</Project>
       <Name>Localization</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\LanguageIntegrator\LanguageIntegrator.csproj">
-      <Project>{a48d9069-d569-4110-9252-a10f139b669b}</Project>
-      <Name>LanguageIntegrator</Name>
     </ProjectReference>
     <ProjectReference Include="..\ParserTools\ParserTools.csproj">
       <Project>{AF2EFD7B-69DD-4B43-AF65-B59B29349C23}</Project>
